+++ packages/benchsdk/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @benchsdk/client
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [1f3555b]
+  - @benchsdk/api@0.5.0
+  - @benchsdk/worker@0.2.4
+
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [43eb478]
+  - @benchsdk/api@0.4.1
+  - @benchsdk/worker@0.2.3
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [6f37302]
+  - @benchsdk/api@0.4.0
+  - @benchsdk/worker@0.2.2
+
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [0b1916f]
+  - @benchsdk/api@0.3.0
+  - @benchsdk/worker@0.2.1
+
 ## 0.4.0
 
 ### Minor Changes